@@ -1,3 +1,4 @@
+publish_to: 'none'
 name: example
 description: A new Flutter project.
 
@@ -14,7 +15,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -30,7 +31,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  device_preview: ^0.5.5
 
 
 # For information on the generic Dart part of this file, see the
