@@ -1,13 +1,13 @@
 import 'package:flutter/cupertino.dart';
 
 class DecorationItem {
-  final Widget decoration;
-  final DateTime date;
-  final Alignment decorationAlignment;
+  final Widget? decoration;
+  final DateTime? date;
+  final Alignment? decorationAlignment;
 
   DecorationItem({
     this.decoration,
     this.date,
     this.decorationAlignment = FractionalOffset.center,
-  });
+  }) : assert(date != null);
 }
